abstract some atoi stuff

+++ libft/src/f_math/ft_log10.c
@@ -6,13 +6,13 @@
 /*   By: jpriou <jpriou@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/12 07:42:12 by jpriou            #+#    #+#             */
-/*   Updated: 2018/08/02 17:30:33 by jpriou           ###   ########.fr       */
+/*   Updated: 2018/08/24 22:22:55 by jpriou           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-double		ft_log10(double n)
+int			ft_log10(double n)
 {
 	return (ft_logb(n, 10));
 }
